+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Added a `foreign_key` setting on model properties for overriding the default foreign key on relationships.
 - Added a `pivot_tablename` setting on model properties for overriding the default pivot table name on belongs-to-many relationships.
 - Added a Collection class to represent a collection of models and provide functionality around managing that collection.
-- Implemented eager loading for has-one relationships.
+- Implemented eager loading for has-one and has-many relationships.
 
 ## Changed
 - Reduce a loadModel call by caching the values after a save.